move external_code
